MB-10496 [4/6]: Add epoch field to Blob; use as part of defragmenter policy
When a Blob object is allocated record the current 'epoch' in it -
essentially how the number of defragmenter passes which have occurred
mod 256. This can then be used to determine the age of a blob, and
consequently can be used by the defragmenter to select Blob objects
which have been in existance (and haven't changed size) for a long
period, and hence are candiates for defragmentation.

Change-Id: I6380e6601ae5ee86ca630031e4edea9711a88433