test_dcp_producer_stream_req_disk: fix occasional unit test failure
This test seems to occasionally fail on factory commit-validation
build - it appears that the number of checkpoints created occasionally
exceeds the expected value; as if we haven't managed to check the
count quickly enough.

Changing the test to check for at least that many checkpoints, instead
of the exact number.

Change-Id: Icb4e4d8b78ca0ed0c9db9e4b83bd3d54b673659a