MB-13591: Defragmenter: ensure time_per_item != 0
During defragmentation we calculate the rate at which items have been
visited, to be able to determine when to pause the defragmenter
task.

However there was a potential divide-by-zero error when calcuating the
time_per_item; if we calculate this as less than one nanosecond it
results in a value of zero, which is used later as divisor.  While one
would assume it would be impossible to process an item in less than
one nanosecond, due to timer precision we may get such a result.

Solution is to add on the timer period to time_per_item; which will
ensure that it is always non-zero.

Change-Id: I6ae26d858c4b4ef4ea4bcc5bd3a108ce18b73268