Use CTest: absolute path to ep.so & ep_{perf,test}suite.so
When specifying the location of the engine and testsuite shared
objects to load, use the absolute path to the files in
CMAKE_CURRENT_BINARY_DIR (i.e. build/ep-engine). This ensures that the
most recent shared object in build is selected, and not a stale one
from install which will be outdated if `make install` hasn't been
recently run.

In other words, `make && make test` in build/ep-engine will use the
correct shared objects.

Change-Id: I92f33d1813a0ff976e8a5353cc36eacf65d85ff5