Regression / unit test for MB-19695
Regression / unit test for MB-19695. See MB for full background of the
problem.

Given the issue is triggered by an unfortunate Task execution order,
creating a unit test for it requires that we can explicitly control
when Tasks and scheduled and executed.

To do this a FakeExecutorPool has been created. It maintains the same
set of Tasks on taskqueues, but no ExecutorThreads exist. Tasks can be
manually fetched from the queues, and then manually run on a
FakeExecutorThread (just runs the Task on the main thread). This
allows unit tests to have complete control on when Tasks are woken up
and run.

Change-Id: I6ef6e051a633c799740aa948eb7531974f94a949