Backport ep-engine_unit_tests from watson to 3.0.x
In Watson we have created a set of 'unit' (i.e. class-level) tests for
ep-engine. To assist in backporting bug fixes, and specifically their
unit tests (to demonstrate they are correct), this patch backports the
test infrastructure itself, along with regression tests for Mb-19503.

Note these tests require GTest, so the CMake changes necessary for it
have also been included.

Tests are a backport from couchbase/watson as of commit
feda304. Modified to handle changes in APIs etc, and to remove tests
which fail on 3.0.x ass we never chose to fix them in the 3.0.x
branch. It also doesn't include the tests in
evp_store_single_threaded_test.cc, as these require changes to
ExecutorPool and friends which haven't been backported.

Change-Id: Iaaf59b0d8d6ba0a2211b630ba00fd837ca01614a