[Ephemeral] Add 'null' EphemeralBucket class and ephemeral test variants
Add a new EphemeralBucket class (c.f. EPBucket), which is instantiated
if 'bucket_type=ephemeral' is set in the configuration. Add variants
of the ep-engine testsuites which run with this bucket_type.

Note: This is initially a placeholder - EphemeralBucket is a subclass
of EPBucket which doesn't override anything, so it's identical to the
Couchstore bucket type in all but name. Subsequent changes will evolve
away from this, but this allows us to start with a complete set of
passing tests.

Change-Id: I4d53550ee589c2e5bea3531e4676cf0226c2fa84