[Ephemeral Bucket]: Refactor seqno generation and ht.unlocked_set()
Currently we generate seqno in checkpoint. Seqno is for a partition
(vbucket) and there may be other modules, other than checkpoint, in the
vbucket that expect a monotonically increasing order of items (for example
if we decide to store an in-memory sequence of items permanently).
Hence generating sequence number outside the checkpoint helps us
develop those modules independent of the checkpoint logic.

Further, this commit refactors the unlocked_set() in hash_table which is
currently loaded with stuff that is not actually related to adding item
to the hash table.

This is code refactor for only set, setWithMeta, addTAPBackfillItem
and replace. Refactoring for delete, item expiry etc will be done
later.

Change-Id: Iac66eef62ee370fda7f297f6d2ab6bc2ca6c9db4