[Ephemeral Bucket]: Add a class that stores ordered sequence of items in memory
This commit adds an abstract base class 'SequenceList' that provides apis
to store ordered sequence of items in memory.

'BasicLinkedList' is a concrete class derived from 'SequenceList' that
implements in-memory ordered seq of items stored in our multi-threaded,
monotonically increasing seq model. 'BasicLinkedList' uses a boost intrusive
doubly linked list. It is essentially a linked list of 'StoredValues'.

'BasicLinkedList' works in tandem with already existing 'HashTable'
to achieve O(1) writes/updates and O(1) reads. Range read is also
implemented which is necessary for an in-memory DCP backfill. Range reads
provide point-in-time snapshots which is essential for our incremental DCP.

'BasicLinkedList' is NOT heavily multi-threaded. Writer threads and
'cleanStaleItems' threads block each other. (Stale items are a consequence
of point-in-time snapshots).

This is still a standalone module yet to be integrated with ep-engine.

Finally a module test 'ephemeral_vb_test' has been added as a module test.

Note: This is still not fully integrated with the kv-engine.

Change-Id: I31bde3ea25509504fa8cc23a5506e7b14fa16bc8