[1/4] Ephemeral class shuffle: Move kvbucket.h to kv_bucket_iface.h
The current class hierarchy for Bucket classes in ep-engine looks
like:

                   KVBucket    - pure interface
                      ^
                      |
                   EPBucket    - concrete, all code

To add support for Ephemeral buckets, we want to create a new
EphemeralBucket class. This is expected to share a majority of its
code with EPBucket (VBuckets, DCP, rebalance, failover), but with
significant changes (no storage, HashTable is ordered).

As such, we want to be able to share the majority of the current
EPBucket's code, but also deviate where necessary. This suggests an
inheritance heirarchy of:

                KVBucketIface  - pure interface
                     ^
                     |
                  KVBucket     - shared code
                     ^
                     |
       /----------------------------\
       |                            |
    EPBucket                 EphemeralBucket

Where each subclass contains the class-specific code, and common code
present in the parent class.

However, it is desirable to maintain as much history in Git as
possible - Git has poor handling of moves of individual pieces of
code, and if we simply created a new shared parent class - KVBucket -
and moved big chunks of code from EPBucket into KVBucket we'd likely
cause many merge conflicts in future.

It is preferable to instead break the change into a series of smaller
patches, where we rename complete files in one patch (which Git
handles very well), and then rename classes (i.e. make code changes)
in a seperate patch.

Therefore this patch is the first of a series of four:

1. Move kvbucket.h to kv_bucket_iface.h (this patch) - change no code,
   but move the file which is logically the abstract base class to the
   correct filename.

2. Rename class KVBucket to KVBucketIface - "make room" in the class
   hierarchy for a new concrete KVBucket class.

3. Move ep.{ccm,h} -> kv_bucket.{cc,h} - change no code, but move the
   file which is to become the new shared parent class to the correct
   filename.

4. Rename class EPBucket -> KVBucket - Finally rename the code which
   is the shared parent class to the correct name.

Change-Id: Ifc06785b3c32a4ccc79f903c1fd37d476b879fd9