[Ephemeral] Add persistence flag to CheckpointConfig XXX TEST FAILURES
Add a new boolean flag to CheckpointConfig - persistenceEnabled. This
is true for EP buckets, but is false for Ephemeral buckets.

This is used for controlling CheckpointManager behaviour which is
specific to the persistence mode - for example if the persistence
cursor should be registered.

NOTE: Currently fails ephemeral tests, they have
      CheckpointConfig::persistence disabled, but still run the
      Flusher task...

Change-Id: I565a8ac5751fd418320f22f6699b34a9bcacbbd1