Remove unnecessary RCPtr<> from CheckpointManager methods
There are a number of CheckpointManager methods which take parameters
via RCPtr<VBucket> where the method doesn't need a ref-counted pointer
- the lifetime of the object is not extended or changed in any way.

As per C++ Core Guidelines[1] R.30, don't use a smart pointer in these
cases. This also makes it easier for subclasses of VBucket
(e.g. EphemeralVBucket) to be passed to these methods.

[1]: http://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rr-smartptrparam

Change-Id: I5c56745924fa7590be8ee2b24b6e05228e543e69