CheckpointTest: Add a ThreadGate class
Add a ThreadGate class - an object which is used to synchronise the
execution of a number of threads.  Each thread calls thread_up(), and
until all thread have called this they are all blocked.

Note: This functionality already existed in the test but used static
variables and hence only allowed it to be used once per program
execution. A forthcoming change will want to run this test twice (with
different parameters), hence this refactoring.

Change-Id: Ic9d2da2369aa553f22d4a336e5c341a7f502687e