KVShard: Only create Flusher for EPBucket
The Flusher is only applicable for EPBuckets, so KVShard should only
crease it in that instance.

Fix a bug in the checking of the object type (we were incorrectly
checking if the bucket was the base class shared with with EP and
Ephemeral), and moreover a dynamic_cast<> is not safe while one is
still in the context of an object's constructor. Instead, just use the
configuration to determine if the bucket is persistent.

Then update a number of places where the flusher was unconditionally
accessed.

Change-Id: I594089b147b8b08672fb53df8fc41f6bdca38d31